setupu typescript and converted loading inditor in tsx
