+++ README.md
@@ -1,28 +1,12 @@
-# [React Mentions](https://react-mentions.vercel.app)
+This is a forked version of [react-mentions](https://github.com/signavio/react-mentions) package.
 
-A React component that let's you mention people in a textarea like you are used to on Facebook or Twitter.
-
-Used in production at [Signavio](https://signavio.com), [State](https://state.com), [Snips](https://snips.ai), [Swat.io](https://swat.io), [GotDone](https://www.gotdone.me), [Volinspire](https://volinspire.com), [Marvin](https://amazingmarvin.com), [Timely](https://timelyapp.com), [GuideFitter](https://www.guidefitter.com/), [Evite](https://www.evite.com/), [Publer](https://publer.me/), [Kontentino](https://www.kontentino.com/), [Wix.com](https://wix.com), [Highlight](https://highlight.run/) and [you?](https://github.com/signavio/react-mentions/edit/master/README.md)
-
-## Getting started
-
-Install the _react-mentions_ package via npm:
-
-```
-npm install react-mentions --save
-```
-
-Or yarn:
-
-```
-yarn add react-mentions
-```
+````
 
 The package exports two React components for rendering the mentions textarea:
 
 ```javascript
 import { MentionsInput, Mention } from 'react-mentions'
-```
+````
 
 `MentionsInput` is the main component rendering the textarea control. It takes one or multiple `Mention` components as its children. Each `Mention` component represents a data source for a specific class of mentionable objects, such as users, template variables, issues, etc.
