+++ package.json
@@ -87,9 +87,14 @@
   },
   "dependencies": {
     "@babel/runtime": "7.4.5",
+    "@types/jest": "^29.5.2",
+    "@types/node": "^20.3.1",
+    "@types/react": "^18.2.13",
+    "@types/react-dom": "^18.2.6",
     "invariant": "^2.2.4",
     "prop-types": "^15.5.8",
-    "substyle": "^9.1.0"
+    "substyle": "^9.1.0",
+    "typescript": "^5.1.3"
   },
   "peerDependencies": {
     "react": ">=16.8.3",
