+++ src/TypescriptCamp/LoadingIndicator/style.ts
@@ -1,0 +1,10 @@
+export const lineStyle = {
+  borderRadius: '1.25rem',
+  height: '0.5rem',
+  marginBottom: '0.5rem',
+  background:
+    'linear-gradient(to right, #99A0A3 0%, #707679 20%, #464A4D 40%, #464A4D 60%, #707679 80% , #99A0A3 100%)',
+  backgroundSize: '1000px',
+  animation: 'placeholderShimmer 1.2s infinite linear',
+  amimationFillMode: 'forwards',
+}
