+++ src/TypescriptCamp/LoadingIndicator/types.ts
@@ -1,0 +1,3 @@
+export interface LoadingIndictorProps {
+  
+}
